requirements.txt
beautifulsoup4==4.9.3
google==3.0.0
grpcio==1.33.1
grpcio-tools==1.33.1
protobuf==3.13.0
six==1.15.0
soupsieve==2.0.1
